fix(integrations): accept float totalDistance in Suunto workout schema
Suunto's API returns totalDistance (meters) as a float, e.g. 21381.4,
but WorkoutJSON typed it as int. Pydantic v2 rejects a float with a
fractional part (int_from_float), so one such workout raised
ValidationError and, since load_data parses the whole payload in a
single list comprehension, aborted the entire sync with an HTTP 500.

Type the field float | None to match its already-float siblings
(maxSpeed, avgSpeed, totalAscent, ...). Downstream is unaffected:
_build_metrics already builds the value via
Decimal(str(raw_workout.totalDistance)).

Adds a regression test that parses a fractional distance and verifies
the resulting Decimal metric.

+++ backend/app/schemas/providers/suunto/workout_import.py
@@ -43,7 +43,10 @@ class WorkoutJSON(BaseModel):
     timeOffsetInMinutes: int | None = None
 
     # Metrics (all optional)
-    totalDistance: int | None = None
+    # Suunto reports distance in meters as a float (e.g. 21381.4); keep it float so
+    # fractional values don't fail Pydantic's int_from_float validation. Downstream
+    # consumers already store it as Decimal(str(...)).
+    totalDistance: float | None = None
     stepCount: int | None = None
     energyConsumption: int | None = None
